pyproject.toml
[project]
name = "axion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Aleš Kučera", email = "kuceral4@fel.cvut.cz" }]
requires-python = ">=3.12"
dependencies = []

[project.optional-dependencies]
sim = [
  "h5py==3.14.0",
  "hydra-core==1.3.2",
  "matplotlib==3.10.3",
  "newton[examples,importers,torch-cu12]",
  "nvtx==0.2.12",
  "openmesh==1.2.1",
  "pandas==2.3.1",
  "plotly==6.4.0",
  "pytest==8.4.0",
  "scipy==1.16.1",
  "taichi==1.7.3",
  "torch>=2.9.1",
  "tqdm==4.67.1",
  "trimesh==4.6.12",
  "usd-core==25.5.1",
  "mujoco-warp>=3.8.0.1",
  "warp-lang>=1.13.0",
]
brax = [
  "brax",
  "jax[cuda12]",
]
genesis = [
  "genesis-world",
]
nimble = [
  "nimblephysics",
]
dev = [
  "axion[sim]",
  "axion[docs]",
]
docs = [
  "mkdocs-gen-files==0.5.0",
  "mkdocs-literate-nav==0.6.2",
  "mkdocs-material==9.6.21",
  "mkdocs-section-index==0.3.10",
  "mkdocstrings-python==1.18.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/axion", "examples"]

[tool.uv.workspace]
members = ["third_party/newton"]

[tool.uv]
conflicts = [
    [
        { package = "newton", extra = "torch-cu12" },
        { package = "newton", extra = "torch-cu13" },
    ],
]

[tool.uv.sources]
newton = { workspace = true }
warp-lang = { index = "nvidia" }

[[tool.uv.index]]
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "nvidia"
url = "https://pypi.nvidia.com/"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
